refactor(payment): make PaymentDate public and change AmountPaid to decimal in entity and DTOs

+++ Sireen.Application/DTOs/Payments/CreatePaymentDto.cs
@@ -12,7 +12,7 @@ namespace Sireen.Application.DTOs.Payments
     public class CreatePaymentDto
     {
         [AmountPaidRequiredForImmediateMethodsAttribute]
-        public double AmountPaid { get; set; }
+        public decimal AmountPaid { get; set; }
 
         [Required(ErrorMessage = "Payment method is required")]
         public PaymentMethod PaymentMethod { get; set; }
