+++ Sireen.Application/DTOs/Payments/PaymentDisplayForBookingDto.cs
@@ -10,7 +10,7 @@ namespace Sireen.Application.DTOs.Payments
     public class PaymentDisplayForBookingDto
     {
         public int Id { get; set; }
-        public double AmountPaid { get; set; }
+        public decimal AmountPaid { get; set; }
         public PaymentStatus PaymentStatus { get; set; }
         public PaymentMethod PaymentMethod { get; set; }
         public DateTime PaymentDate { get; set; } = DateTime.UtcNow;
